+++ yampt.editor/source/view/nav_tree_view.cpp
@@ -106,6 +106,11 @@ void nav_tree_view_t::set_patch_plugins(const std::set<std::string> * patch)
 	m_model->set_patch_plugins(patch);
 }
 
+void nav_tree_view_t::set_dirty_plugins(const std::set<std::string> * dirty)
+{
+	m_model->set_dirty_plugins(dirty);
+}
+
 void nav_tree_view_t::set_editable_columns(const editable_column_set_t * editable)
 {
 	m_model->set_editable_columns(editable);