+++ yampt.editor/source/view/nav_tree_view.hpp
@@ -25,6 +25,7 @@ public:
 	void set_show_deleted_strikeout(bool value);
 	void set_excluded_plugins(const std::set<std::string> * excluded);
 	void set_patch_plugins(const std::set<std::string> * patch);
+	void set_dirty_plugins(const std::set<std::string> * dirty);
 	void set_editable_columns(const editable_column_set_t * editable);
 	void set_display_codepage(codepage_t codepage);